// src/api-types.ts
export namespace CustomTypes {
  export type RichText = {
    rich_text: Array<RichTextItemRequest>;
    type?: "rich_text";
  };
  export type Title = {
    title: Array<RichTextItemRequest>;
    type?: "title";
  };
  export type Number = {
    number: number | null;
    type?: "number";
  };
  export type Date = {
    date: DateRequest | null;
    type?: "date";
  };
  export type Select = {
    select:
      | {
          id: StringRequest;
          name?: StringRequest;
          color?: SelectColor;
        }
      | null
      | {
          name: StringRequest;
          id?: StringRequest;
          color?: SelectColor;
        }
      | null;
    type?: "select";
  };
  export type MultiSelect = {
    multi_select: Array<
      | {
          id: StringRequest;
          name?: StringRequest;
          color?: SelectColor;
        }
      | {
          name: StringRequest;
          id?: StringRequest;
          color?: SelectColor;
        }
    >;
    type?: "multi_select";
  };
  export type URL = {
    url: TextRequest | null;
    type?: "url";
  };
}

export type RichTextItemRequest =
  | {
      text: {
        content: string;
        link?: {
          url: TextRequest;
        } | null;
      };
      type?: "text";
      annotations?: {
        bold?: boolean;
        italic?: boolean;
        strikethrough?: boolean;
        underline?: boolean;
        code?: boolean;
        color?:
          | "default"
          | "gray"
          | "brown"
          | "orange"
          | "yellow"
          | "green"
          | "blue"
          | "purple"
          | "pink"
          | "red"
          | "gray_background"
          | "brown_background"
          | "orange_background"
          | "yellow_background"
          | "green_background"
          | "blue_background"
          | "purple_background"
          | "pink_background"
          | "red_background";
      };
    }
  | {
      mention:
        | {
            user:
              | {
                  id: IdRequest;
                }
              | {
                  person: {
                    email?: string;
                  };
                  id: IdRequest;
                  type?: "person";
                  name?: string | null;
                  avatar_url?: string | null;
                  object?: "user";
                }
              | {
                  bot:
                    | EmptyObject
                    | {
                        owner:
                          | {
                              type: "user";
                              user:
                                | {
                                    type: "person";
                                    person: {
                                      email: string;
                                    };
                                    name: string | null;
                                    avatar_url: string | null;
                                    id: IdRequest;
                                    object: "user";
                                  }
                                | {
                                    id: IdRequest;
                                    object: "user";
                                  };
                            }
                          | {
                              type: "workspace";
                              workspace: true;
                            };
                      };
                  id: IdRequest;
                  type?: "bot";
                  name?: string | null;
                  avatar_url?: string | null;
                  object?: "user";
                };
          }
        | {
            date: DateRequest;
          }
        | {
            page: {
              id: IdRequest;
            };
          }
        | {
            database: {
              id: IdRequest;
            };
          };
      type?: "mention";
      annotations?: {
        bold?: boolean;
        italic?: boolean;
        strikethrough?: boolean;
        underline?: boolean;
        code?: boolean;
        color?:
          | "default"
          | "gray"
          | "brown"
          | "orange"
          | "yellow"
          | "green"
          | "blue"
          | "purple"
          | "pink"
          | "red"
          | "gray_background"
          | "brown_background"
          | "orange_background"
          | "yellow_background"
          | "green_background"
          | "blue_background"
          | "purple_background"
          | "pink_background"
          | "red_background";
      };
    }
  | {
      equation: {
        expression: TextRequest;
      };
      type?: "equation";
      annotations?: {
        bold?: boolean;
        italic?: boolean;
        strikethrough?: boolean;
        underline?: boolean;
        code?: boolean;
        color?:
          | "default"
          | "gray"
          | "brown"
          | "orange"
          | "yellow"
          | "green"
          | "blue"
          | "purple"
          | "pink"
          | "red"
          | "gray_background"
          | "brown_background"
          | "orange_background"
          | "yellow_background"
          | "green_background"
          | "blue_background"
          | "purple_background"
          | "pink_background"
          | "red_background";
      };
    };
// ...
export type DateRequest = {
  start: string;
  end?: string | null;
  time_zone?: TimeZoneRequest | null;
};

declare type DateResponse = {
  start: string;
  end: string | null;
  time_zone: TimeZoneRequest | null;
};

export type StringRequest = string;

export type SelectColor =
  | "default"
  | "gray"
  | "brown"
  | "orange"
  | "yellow"
  | "green"
  | "blue"
  | "purple"
  | "pink"
  | "red";

export type IdRequest = string | string;

export type EmptyObject = Record<string, never>;

export type TextRequest = string;